chore(refactor): Renamed components
