+++ app/components/Evidence/Evidenceform/Evidenceform.tsx
@@ -1,0 +1,89 @@
+'use client';
+
+import { evidenceSchema } from '@/app/schemas';
+import { Creator, EvidenceFromDto } from '@/app/types';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { ChangeEvent, FC, useState } from 'react';
+import { useForm } from 'react-hook-form';
+import CreatorsList from '../formSections/creatorInput';
+import CustomDatePicker from '../formSections/dateOfCreation';
+import FileUpload from '../formSections/fileUpload';
+import FileUploadForLicense from '../formSections/fileUploadforLicence';
+import LicenseSelect from '../formSections/licenseSelect';
+import Medium from '../formSections/medium';
+import WorkNameInput from '../formSections/workNameInpute';
+
+import './Evidenceform.scss';
+
+const Evidenceform: FC = () => {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<EvidenceFromDto>({
+    resolver: zodResolver(evidenceSchema),
+    defaultValues: {},
+  });
+
+  const [selectedDate, setSelectedDate] = useState<Date | null>(null);
+  const [creators, setCreators] = useState<Creator[]>([{ id: 1, name: '' }]);
+  const [selectedLicense, setSelectedLicense] = useState<string>('');
+
+  const handleFileUpload = (e: ChangeEvent<HTMLInputElement>) => {
+    // setFile(e.target.files[0]);
+  };
+
+  const addCreator = () => {
+    const newId = creators.length + 1;
+    setCreators([...creators, { id: newId, name: '' }]);
+  };
+
+  const handleNameChange = (id: number, value: string) => {
+    setCreators((prevCreators) =>
+      prevCreators.map((creator) =>
+        creator.id === id ? { ...creator, name: value } : creator
+      )
+    );
+  };
+
+  const handleLicenseChange = (event: ChangeEvent<HTMLSelectElement>) => {
+    setSelectedLicense(event.target.value);
+  };
+
+  const handleLicenseUpload = (e: ChangeEvent<HTMLInputElement>) => {
+    // Your file upload logic here
+  };
+
+  return (
+    <div className='p-5 lg:p-10'>
+      <h3 className='pb-4 text-4xl font-bold tracking-tight text-gray-500'>
+        Evidence Form
+      </h3>
+      <form className='w-[100%] space-y-6 md:w-[328px]'>
+        <CreatorsList
+          creators={creators}
+          onAddCreator={addCreator}
+          onNameChange={handleNameChange}
+        />
+        <WorkNameInput />
+        <Medium />
+        <FileUpload handleFileUpload={handleFileUpload} />
+        <CustomDatePicker
+          selectedDate={selectedDate}
+          onChange={setSelectedDate}
+        />
+        <LicenseSelect
+          selectedLicense={selectedLicense}
+          handleLicenseChange={handleLicenseChange}
+          handleLicenseUpload={handleFileUpload}
+        />
+        <FileUploadForLicense
+          handleLicenseUpload={handleLicenseUpload}
+          selectedLicense={selectedLicense}
+        />
+        <button className='btn-primary btn w-full'>Submit</button>
+      </form>
+    </div>
+  );
+};
+export default Evidenceform;
