+++ app/schemas/index.ts
@@ -1,0 +1,3 @@
+import evidenceSchema from './evidenceSchema';
+
+export { evidenceSchema };
