+++ app/types/evidence.ts
@@ -1,0 +1,13 @@
+export type Creator = {
+  id: number;
+  name: string;
+};
+
+export type EvidenceFromDto = {
+  creators: Creator[];
+  nameOfWork: string;
+  medium: string;
+  work: File;
+  dateOfCreation: Date;
+  license: string;
+};
